Continue warning clean up
